Переименовал js файлы
