+++ src/view/info-main.js
